+++ .vscode/settings.json
@@ -1,0 +1,6 @@
+{
+    "files.associations": {
+        "complex": "cpp",
+        "functional": "cpp"
+    }
+}
