--- px4_utils/package.xml
@@ -1,40 +1,0 @@
-<?xml version="1.0"?>
-<package format="2">
-  <name>px4_utils</name>
-  <version>0.0.0</version>
-  <description>The px4_utils package</description>
-
-  <maintainer email="jaimefriedhelmzhao@gmail.com">Zhaohong Liu</maintainer>
-
-
-  <!-- One license tag required, multiple allowed, one license per tag -->
-  <!-- Commonly used license strings: -->
-  <!--   BSD, MIT, Boost Software License, GPLv2, GPLv3, LGPLv2.1, LGPLv3 -->
-  <license>TODO</license>
-
-  <buildtool_depend>catkin</buildtool_depend>
-
-  <build_depend>mavros_msgs</build_depend>
-  <build_depend>roscpp</build_depend>
-  <build_depend>rospy</build_depend>
-  <build_depend>std_msgs</build_depend>
-  <build_depend>quadrotor_msgs</build_depend>
-
-  <build_export_depend>mavros_msgs</build_export_depend>
-  <build_export_depend>roscpp</build_export_depend>
-  <build_export_depend>rospy</build_export_depend>
-  <build_export_depend>std_msgs</build_export_depend>
-
-  <exec_depend>mavros_msgs</exec_depend>
-  <exec_depend>roscpp</exec_depend>
-  <exec_depend>rospy</exec_depend>
-  <exec_depend>std_msgs</exec_depend>
-  <exec_depend>quadrotor_msgs</exec_depend>
-
-
-  <!-- The export tag contains other, unspecified, tags -->
-  <export>
-    <!-- Other tools can request additional information be placed here -->
-
-  </export>
-</package>
