+++ px4_utils_land/launch/dev_px4_fsm_interface.launch
@@ -1,0 +1,15 @@
+<launch>
+    <node pkg="px4_utils_land" type="px4_ctrl_fsm_node_land" name="px4_ctrl_fsm" output="screen">
+        <param name="px4fsm/exec_period" value="0.01" type="double"/>
+        <param name="px4fsm/cruise_height" value="2.0" type="double"/>
+        <param name="px4fsm/fence_x" value="20.0" type="double"/>
+        <param name="px4fsm/fence_y" value="20.0" type="double"/>
+        <param name="px4fsm/fence_z" value="10.0" type="double"/>
+        <param name="px4fsm/land_cmd_topic" value="/trigger_landing" />
+	<param name="image_topic" value="/camera/color/image_raw" type="string" />
+    </node>
+
+    <node pkg="px4_utils_land" type="px4_fsm_interface_node_land" name="px4_fsm_interface" output="screen">
+        <param name="land_topic" value="/trigger_landing" />
+    </node>
+</launch>
