+++ px4_utils_land/scripts/nomachine_headless_check.sh
